requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
configparse==0.1.5
DateTime==4.4
dnspython==2.2.1
idna==3.3
panda==0.3.1
pyaes==1.6.1
pyasn1==0.4.8
pymongo==4.3.2
pyTelegramBotAPI==4.5.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.9
six==1.16.0
Telethon==1.25.4
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wincertstore==0.2
zope.interface==5.4.0
